[DRG] Correct Dragoon Wyvern weapon damage to match data analysis
The following applies up to level 75 to our knowledge.

Per Jimmayus:

Previously, Wyvern weapon damage was generated using a system with no apparent source.
This value was arbitrarily higher than anecdotal observations, but until recently we did not have the required parse data to accurately derive a formula with acceptable precision. Casey recently sourced an acceptably high sample of parse data for wyverns on a static set of targets which allowed us to generate meaningful conclusions from the data.

The test set were Robber Crabs of levels 61-63, whose relevant VIT values and DEF values according to /check method reconfirmations indicate that Wyvern base damage and susceptibility to Boolean value observation (the so-called "1.0 PDIF spike" observable in player data and as documented in the formula found in https://ffxiclopedia.fandom.com/wiki/Level_Correction_Function_and_pDIF#Distribution_of_melee_hits ) would be unaffected by differences on monster level.
As wyverns are npcs level correction is a factor, and the results indicate a PDIF spike when Crab defense buffs were present (which is predicted by the formula).
The distribution of wyvern hits is observable here: https://docs.google.com/spreadsheets/d/1YBoveP-weMdidrirY-vPDzHyxbEI2ryECINlfCnFkLI/edit?usp=sharing.
As noted above the PDIF spike is present at the lower range, and our model would predict its existence only when a Defense Buff is present.
The range is arbitrarily larger than secondary randomizer would predict, but this is consistent with STR down debuffs being applied to the wyvern (unfortunately this isn't filterable, but the range present is consistent with decayed debuff effects).

The conclusions are as follows:

1) Wyverns benefit positively from level correction, and have STR and Attack values consistent with estimates based on other monster and pet STR and Attack calculations.
2) Wyvern Base Damage is consistent with an offset of 3 and a slope of .5, or D: 3 + floor(level * .5)