[core] Adjust TH handling to match retail
Source: https://www.bluegartr.com/threads/50199
Q: Do multiple instances of Treasure Hunter in a single party increase the chances of a drop? When exactly does the effect of Treasure Hunter take place?
A: When more than one member has Treasure Hunter, the player with the strongest version will have theirs take effect for the party. The effect comes into play as soon as the Thief generates enmity from the target enemy. Even if the Thief dies, the effect remains.